backend/requirements.txt
fastapi[all]
uvicorn
pymysql
sqlalchemy
mysql-connector-python
python-dotenv
requests
aiohttp
google-generativeai
passlib[bcrypt]
pylance
cryptography
python-jose